README.md
In this program, you have to guess a random number by iterations. When you enter each number, it is compared with the guessed one and the console displays the result of the comparison. To complicate the game, a timer has been added here. The timer time is also up to you.